demo-app/services/checkout/requirements.txt
# Runtime dependencies for the checkout service.
# Pinned to known-good versions; bump deliberately, not opportunistically.

fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3

# OpenTelemetry — auto-instrumentation for FastAPI + requests + manual API.
# Versions matched to opentelemetry-operator-autoinstrumentation-python image.
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-semantic-conventions==0.50b0
